parse_size: rewrite to simplify & handle more edge cases
Rewrite the code to avoid checking __WORDSIZE, and to simplify the
overflow logic.  Add more testcases that the old code didn't handle
correctly.  The API is now consistently 64-bit, and moves the value
too big checking to the caller when it wants a smaller sizes (like
size_t on 32-bit systems).  It also compiles smaller & faster since
it doesn't rely on division at all, and there's only 1 multiplication
when a suffix is actually specified.

This rewrite also fixes an, albeit harmless, integer underflow as
found by coverity.  The last iteration of the `while (i-- > 0)` loop
will make the unsigned i wrap backaround to SIZE_MAX.

BUG=None
TEST=CQ passes

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minijail/+/5656294
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>

+++ util.h
@@ -10,6 +10,7 @@
 #define _UTIL_H_
 
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -259,7 +260,19 @@ long int parse_single_constant(char *constant_str, char **endptr)
     attribute_nonnull();
 long int parse_constant(char *constant_str, char **endptr)
     attribute_nonnull((1));
-int parse_size(size_t *size, const char *sizespec) attribute_nonnull();
+
+/*
+ * parse_size: parse a string to a positive integer bytes with optional suffix.
+ * @size     The output parsed size, in bytes
+ * @sizespec The input string to parse
+ *
+ * A single 1-char suffix is supported like "10K" or "6G".  These use base 1024,
+ * not base 1000.  i.e. "1K" is "1024".  It is case-sensitive.
+ *
+ * Returns 0 on success, negative errno on failure.
+ * Only writes to |size| on success.
+ */
+int parse_size(uint64_t *size, const char *sizespec) attribute_nonnull();
 
 char *strip(char *s) attribute_nonnull();
